Split traits to their own header
