+++ source/library/reflect/public/potato/reflect/CMakeLists.txt
@@ -5,4 +5,5 @@ target_sources(reflect PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/reflect.h
     ${CMAKE_CURRENT_SOURCE_DIR}/json_serializer.h
     ${CMAKE_CURRENT_SOURCE_DIR}/stream_serializer.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/traits.h
 )
